# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: ShieldBinary API
  description: Binary protection service for .NET and native PE
  version: 1.0.0

servers:
  - url: /
    description: API root
  - url: /api/v1
    description: API v1

paths:
  /health:
    get:
      summary: Liveness check
      tags: [Health]
      responses:
        '200': { description: Server is running }

  /ready:
    get:
      summary: Readiness check (Redis connectivity)
      tags: [Health]
      responses:
        '200': { description: Ready to accept traffic }
        '503': { description: Unhealthy (e.g. Redis down) }

  /auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '200':
          description: Registration successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { description: Email already registered }

  /auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token: { type: string }
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Invalid credentials }

  /auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
        '401': { description: Unauthorized }

  /upload:
    post:
      summary: Upload a binary file
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: .exe or .dll file (max 100MB)
      responses:
        '200':
          description: Upload successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  input_key: { type: string }
                  filename: { type: string }
                  size: { type: integer }
        '400': { $ref: '#/components/responses/BadRequest' }

  /scan:
    post:
      summary: Scan an executable
      description: Scan .exe/.dll for PE type, architecture, and BinaryProtect protection markers. No storage; result returned immediately.
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: .exe or .dll file (max 100MB)
      responses:
        '200':
          description: Scan result
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ScanResult' }
        '400': { $ref: '#/components/responses/BadRequest' }

  /jobs:
    get:
      summary: List jobs for current user
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: List of jobs
          content:
            application/json:
              schema:
                type: object
                properties:
                  jobs:
                    type: array
                    items: { $ref: '#/components/schemas/Job' }
    post:
      summary: Create a protection job
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input_key, tier]
              properties:
                input_key: { type: string }
                tier: { type: string, enum: [minimal, basic, pro, enterprise] }
                binary_type: { type: string, default: auto }
                low_entropy: { type: boolean, default: false }
                polymorphic_mode: { type: boolean, default: false }
                protections:
                  type: array
                  items: { type: string }
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                type: object
                properties:
                  job_id: { type: string }
                  status: { type: string, example: queued }
                  message: { type: string }
    delete:
      summary: Delete all jobs (clear history)
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: All jobs deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  count: { type: integer }
        '401': { description: Unauthorized }

  /jobs/{id}:
    get:
      summary: Get job status
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Job' }
        '404': { description: Job not found }
    delete:
      summary: Delete a job
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Job deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
        '404': { description: Job not found }

  /jobs/{id}/download:
    get:
      summary: Download protected binary (one-time)
      description: Streams the protected file. After a successful download, the file is deleted from storage and cannot be downloaded again.
      tags: [Jobs]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: File or presigned URL
        '400': { description: Job not ready }
        '404': { description: Job not found }

  /jobs/{id}/threat-intel/submit:
    post:
      summary: Submit protected output to threat intelligence provider (manual opt-in)
      tags: [ThreatIntel]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Already submitted or immediate verdict available
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThreatIntelJobStatus' }
        '202':
          description: Submitted and pending provider analysis
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThreatIntelJobStatus' }
        '400': { description: Bad request }
        '404': { description: Job not found }

  /jobs/{id}/threat-intel:
    get:
      summary: Get threat intelligence status for a job
      tags: [ThreatIntel]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Threat intelligence status for this job
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ThreatIntelJobStatus' }
        '404': { description: Job not found }

  /threat-intel/flags:
    get:
      summary: List technique detection flags
      tags: [ThreatIntel]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 50 }
      responses:
        '200':
          description: Flag list
          content:
            application/json:
              schema:
                type: object
                properties:
                  enabled: { type: boolean }
                  flags:
                    type: array
                    items: { $ref: '#/components/schemas/TechniqueFlag' }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string }

    ScanResult:
      type: object
      properties:
        path: { type: string }
        size: { type: integer }
        valid_pe: { type: boolean }
        is_dotnet: { type: boolean }
        machine: { type: string }
        protected: { type: boolean }
        protection_format: { type: string }
        protection_tier: { type: string }
        payload_size: { type: integer }
        embedding: { type: string }
        error: { type: string }

    Job:
      type: object
      properties:
        job_id: { type: string }
        status: { type: string }
        progress: { type: integer }
        tier: { type: string }
        binary_type: { type: string }
        low_entropy: { type: boolean }
        polymorphic_mode: { type: boolean }
        protections:
          type: array
          items: { type: string }
        pass_metrics:
          type: array
          items: { $ref: '#/components/schemas/PassMetric' }
        size_impact: { $ref: '#/components/schemas/SizeImpact' }
        compatibility_report: { $ref: '#/components/schemas/CompatibilityReport' }
        strength_score: { $ref: '#/components/schemas/StrengthScore' }
        retry_suggestions:
          type: array
          items: { $ref: '#/components/schemas/RetrySuggestion' }
        input_key: { type: string }
        output_key: { type: string }
        error: { type: string }

    PassMetric:
      type: object
      properties:
        name: { type: string }
        duration_ms: { type: integer, format: int64 }
        success: { type: boolean }
        error: { type: string }
        size_delta_bytes: { type: integer, format: int64 }

    SizeImpact:
      type: object
      properties:
        input_bytes: { type: integer, format: int64 }
        output_bytes: { type: integer, format: int64 }
        pass_deltas:
          type: object
          additionalProperties:
            type: integer
            format: int64

    CompatibilityReport:
      type: object
      properties:
        status: { type: string }
        mode: { type: string }
        exit_code: { type: integer }
        timed_out: { type: boolean }
        stdout_snippet: { type: string }
        stderr_snippet: { type: string }
        notes: { type: string }

    StrengthScore:
      type: object
      properties:
        score: { type: integer, minimum: 0, maximum: 100 }
        band: { type: string }
        time_estimate: { type: string }

    RetrySuggestion:
      type: object
      properties:
        label: { type: string }
        reason: { type: string }
        tier: { type: string }
        low_entropy: { type: boolean }
        polymorphic_mode: { type: boolean }
        protections:
          type: array
          items: { type: string }

    ThreatIntelJobStatus:
      type: object
      properties:
        enabled: { type: boolean }
        submitted: { type: boolean }
        job_id: { type: string }
        sample_hash: { type: string }
        provider: { type: string }
        status: { type: string }
        provider_submission: { type: string }
        analysis_status: { type: string }
        detected_count: { type: integer }
        engine_count: { type: integer }
        verdict_ratio: { type: number, format: float }
        submitted_at: { type: string, format: date-time }
        last_seen: { type: string, format: date-time }
        last_error: { type: string }

    TechniqueFlag:
      type: object
      properties:
        technique_key: { type: string }
        severity: { type: string }
        reason: { type: string }
        opened_at: { type: string, format: date-time }
        closed_at: { type: string, format: date-time }
        state: { type: string }
        last_detected_ratio: { type: number, format: float }
        last_sample_count: { type: integer }

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }